+++ addOns/graaljs/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add info URL.
 
 ### Changed
+- Update minimum ZAP version to 2.13.0.
 - Replace usage of singletons with injected variables (e.g. `model`, `control`) in scripts.
 
 ### Fixed