+++ addOns/graaljs/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [0.4.0] - 2023-07-11
 ### Added
 - Add info URL.
 
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 First version.
 
+[0.4.0]: https://github.com/zaproxy/zap-extensions/releases/graaljs-v0.4.0
 [0.3.0]: https://github.com/zaproxy/zap-extensions/releases/graaljs-v0.3.0
 [0.2.0]: https://github.com/zaproxy/zap-extensions/releases/graaljs-v0.2.0
 [0.1.0]: https://github.com/zaproxy/zap-extensions/releases/graaljs-v0.1.0